Rename `channels` module into `replicon_channels`
Because it contains `RepliconChannels` with just a few enums.
