+++ src/server/replication_messages/mutate_message.rs
@@ -5,12 +5,12 @@ use postcard::experimental::{max_size::MaxSize, serialized_size};
 
 use super::{component_changes::ComponentChanges, serialized_data::SerializedData};
 use crate::core::{
-    channels::ReplicationChannel,
     postcard_utils,
     replication::{
         client_ticks::{ClientTicks, EntityBuffer},
         mutate_index::MutateIndex,
     },
+    replicon_channels::ReplicationChannel,
     replicon_server::RepliconServer,
     replicon_tick::RepliconTick,
 };
